add client to download files from open data server
