+++ src/main/java/org/gridsuite/odre/server/services/OdreServiceImpl.java
@@ -34,7 +34,7 @@ public class OdreServiceImpl implements OdreService {
     private static final String GEO_DATA_API_VERSION = "v1";
 
     @Autowired
-    @Qualifier("odreCsvClientImpl")
+    @Qualifier("odreOpenDataClientImpl")
     private OdreClient client;
 
     private RestTemplate geoDataServerRest;
